requirements.txt
# Core
numpy
pyyaml
scikit-learn
matplotlib

# Torch
torch
torchvision
opacus
